+++ app/scripts/TiledPlot.js
@@ -626,6 +626,8 @@ class TiledPlot extends React.Component {
       tracks,
       forceUpdate: Math.random()
     });
+
+    this.props.onResizeTrack();
   }
 
   closeMenus() {
@@ -2106,11 +2108,13 @@ class TiledPlot extends React.Component {
           disableTrackMenu={this.props.disableTrackMenu}
           dragging={this.props.dragging}
           galleryDim={this.galleryDim}
+          getLockGroupExtrema={this.props.getLockGroupExtrema}
           height={this.state.height}
           initialXDomain={this.props.initialXDomain}
           initialYDomain={this.props.initialYDomain}
           isRangeSelection={this.props.mouseTool === MOUSE_TOOL_SELECT}
           isShowGlobalMousePosition={this.props.isShowGlobalMousePosition}
+          isValueScaleLocked={this.props.isValueScaleLocked}
           leftWidth={this.leftWidth}
           leftWidthNoGallery={this.leftWidthNoGallery}
           metaTracks={this.props.metaTracks}
@@ -2355,9 +2359,11 @@ TiledPlot.propTypes = {
   dragging: PropTypes.bool,
   draggingHappening: PropTypes.bool,
   editable: PropTypes.bool,
+  getLockGroupExtrema: PropTypes.func,
   initialXDomain: PropTypes.array,
   initialYDomain: PropTypes.array,
   isShowGlobalMousePosition: PropTypes.bool,
+  isValueScaleLocked: PropTypes.func,
   marginBottom: PropTypes.number.isRequired,
   marginLeft: PropTypes.number.isRequired,
   marginRight: PropTypes.number.isRequired,
@@ -2384,6 +2390,7 @@ TiledPlot.propTypes = {
   onTracksAdded: PropTypes.func,
   onUnlockValueScale: PropTypes.func,
   onValueScaleChanged: PropTypes.func,
+  onResizeTrack: PropTypes.func,
   overlays: PropTypes.array,
   openModal: PropTypes.func,
   pixiRenderer: PropTypes.object,