+++ app/scripts/configs/tracks-info.js
@@ -171,7 +171,8 @@ export const TRACKS_INFO = [
       'showMousePosition',
       'mousePositionColor',
       'showTooltip',
-      'extent'
+      'extent',
+      'zeroValueColor'
     ],
     defaultOptions: {
       backgroundColor: '#eeeeee',
@@ -200,7 +201,8 @@ export const TRACKS_INFO = [
       showMousePosition: false,
       mousePositionColor: '#000000',
       showTooltip: false,
-      extent: 'full'
+      extent: 'full',
+      zeroValueColor: null
     },
     defaultOptionsByTheme: {
       [THEME_DARK]: {
@@ -611,7 +613,8 @@ export const TRACKS_INFO = [
       'heatmapValueScaling',
       'selectRows',
       'colorbarBackgroundColor',
-      'colorbarPosition'
+      'colorbarPosition',
+      'zeroValueColor'
     ],
     defaultOptions: {
       labelPosition: 'topLeft',
@@ -630,7 +633,8 @@ export const TRACKS_INFO = [
       heatmapValueScaling: 'log',
       selectRows: null,
       colorbarBackgroundColor: '#ffffff',
-      colorbarPosition: 'topRight'
+      colorbarPosition: 'topRight',
+      zeroValueColor: null
     },
     defaultOptionsByTheme: {
       [THEME_DARK]: {
@@ -1157,7 +1161,7 @@ export const TRACKS_INFO = [
     type: 'horizontal-gene-annotations',
     datatype: ['gene-annotation'],
     local: false,
-    defaultHeight: 55,
+    defaultHeight: 90,
     orientation: '1d-horizontal',
     name: 'Gene Annotations',
     thumbnail: svgGeneAnnotationsIcon,
@@ -1199,7 +1203,7 @@ export const TRACKS_INFO = [
       trackBorderColor: 'black',
       showMousePosition: false,
       mousePositionColor: '#000000',
-      geneAnnotationHeight: 12,
+      geneAnnotationHeight: 16,
       geneLabelPosition: 'outside',
       geneStrandSpacing: 4
     },
@@ -1217,7 +1221,7 @@ export const TRACKS_INFO = [
     type: 'vertical-gene-annotations',
     datatype: ['gene-annotation'],
     local: false,
-    defaultWidth: 55,
+    defaultWidth: 90,
     orientation: '1d-vertical',
     name: 'Gene Annotations',
     thumbnail: svgVerticalGeneAnnotationsIcon,
@@ -1238,7 +1242,7 @@ export const TRACKS_INFO = [
       'showMousePosition',
       'mousePositionColor',
       'fontSize',
-      'geneAnnoHeight',
+      'geneAnnotationHeight',
       'geneLabelPosition',
       'geneStrandSpacing'
     ],
@@ -1257,7 +1261,7 @@ export const TRACKS_INFO = [
       trackBorderColor: 'black',
       showMousePosition: false,
       mousePositionColor: '#000000',
-      geneAnnoHeight: 12,
+      geneAnnotationHeight: 16,
       geneLabelPosition: 'outside',
       geneStrandSpacing: 4
     },